refactor the edit form components to set up for enabling caching
